Part1 of item rendering

+++ src/PoECommerce.Client.Model/Model/Search/FilterMagnitude.cs
@@ -2,7 +2,7 @@
 {
     public class FilterMagnitude
     {
-        public int? Min { get; set; }
-        public int? Max { get; set; }
+        public double? Min { get; set; }
+        public double? Max { get; set; }
     }
 }
